# web/src/shared/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentAttemptList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssessmentAttemptV1"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AssessmentAttemptList",
        "type": "object"
      },
...
      "DemandValidationRequest": {
        "properties": {
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "validated_demand": {
            "anyOf": [
              {
                "enum": [
                  "recall",
                  "explain",
                  "compare",
                  "apply",
                  "diagnose",
                  "evaluate",
                  "design"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated Demand"
          },
          "validator_kind": {
            "const": "user",
            "default": "user",
            "title": "Validator Kind",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "validated_demand",
          "rationale"
        ],
        "title": "DemandValidationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TheGrandQuiz Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/learning/attempts": {
      "get": {
        "operationId": "list_assessment_attempts_api_v1_learning_attempts_get",
        "parameters": [
          {
            "in": "query",
            "name": "trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentAttemptList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assessment Attempts",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts/{attempt_id}": {
      "get": {
        "operationId": "get_assessment_attempt_api_v1_learning_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentAttemptV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assessment Attempt",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts/{attempt_id}/demand-validations": {
      "post": {
        "operationId": "validate_attempt_demand_api_v1_learning_attempts__attempt_id__demand_validations_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandValidationV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Attempt Demand",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts/{attempt_id}/verdict-corrections": {
      "post": {
        "operationId": "correct_attempt_verdict_api_v1_learning_attempts__attempt_id__verdict_corrections_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentAttemptV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Attempt Verdict",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/items/{item_id}/classifications": {
      "get": {
        "operationId": "get_knowledge_item_classifications_api_v1_learning_items__item_id__classifications_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeClassificationHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge Item Classifications",
        "tags": [
          "learning"
        ]
      },
      "post": {
        "operationId": "classify_knowledge_item_api_v1_learning_items__item_id__classifications_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeClassificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeClassificationV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify Knowledge Item",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/items/{item_id}/classifications/{classification_id}/review": {
      "post": {
        "operationId": "review_knowledge_classification_api_v1_learning_items__item_id__classifications__classification_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "classification_id",
            "required": true,
            "schema": {
              "title": "Classification Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeClassificationV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Knowledge Classification",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/items/{item_id}/tags": {
      "get": {
        "operationId": "list_managed_tags_api_v1_learning_items__item_id__tags_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagAssignmentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Managed Tags",
        "tags": [
          "learning"
        ]
      },
      "post": {
        "operationId": "assign_managed_tag_api_v1_learning_items__item_id__tags_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagAssignmentV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Managed Tag",
        "tags": [
          "learning"
        ]
      }
    },
...
    "/api/v1/learning/projections/{item_id}": {
      "get": {
        "operationId": "get_learner_projection_api_v1_learning_projections__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learner Projection",
        "tags": [
          "learning"
        ]
      }
    },
...
  }
}